backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
qdrant-client
sentence-transformers
transformers
torch
langchain
pypdf
python-multipart
diskcache